# swagger/simple_api.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pouch API Server",
    "description":
      "Pouch api server. [c8112002/pouch-api](https://github.com/c8112002/pouch-api)",
    "contact": {
      "email": "c8112002@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://api:8080/v1",
      "description": "dev server"
    },
    {
      "url": "https://foo/v1",
      "description": "prod server"
    }
  ],
  "tags": [
    {
      "name": "article",
      "description": "Everything about Articles"
    }
  ],
  "paths": {
    "/articles": {
      "get": {
        "tags": ["article"],
        "summary": "Return articles",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "description": "取得件数",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "first_cursor",
            "in": "query",
            "required": true,
            "description": "カーソル(取得する先頭の記事ID)",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Articles"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/article": {
      "post": {
        "tags": ["article"],
        "summary": "記事を作成する",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string"
                  }
                }
              },
              "examples": {
                "url": {
                  "summary": "article's url",
                  "value": "https://google.com"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/articles/{articleId}/tags": {
      "patch": {
        "tags": ["article"],
        "summary": "記事のタグを編集する",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "description": "article id",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "examples": {
                "tags": {
                  "summary": "article's tags",
                  "valuel": ["tag1", "tag2"]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/articles/{articleId}/comment": {
      "patch": {
        "tags": ["article"],
        "summary": "記事のコメントを編集する",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "description": "article id",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comment": {
                    "type": "string"
                  }
                }
              },
              "examples": {
                "comment": {
                  "summary": "article's comment",
                  "valuel": "#comment1 コメントだよ"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/articles/{articleId}/read": {
      "patch": {
        "tags": ["article"],
        "summary": "記事を既読/未読にする",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "description": "article id",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "read": {
                    "type": "boolean"
                  }
                }
              },
              "examples": {
                "url": {
                  "summary": "article's url",
                  "value": "https://google.com"
                },
                "tags": {
                  "summary": "article's tags",
                  "valuel": ["tag1", "tag2"]
                },
                "comment": {
                  "summary": "article's comment(markdown)",
                  "valuel": "#comment1 コメントだよ"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/articles/{articleId}": {
      "delete": {
        "tags": ["article"],
        "summary": "記事を削除する",
        "responses": {
          "200": {
            "description": "OK"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/tags": {
      "get": {
        "tags": ["tag"],
        "summary": "Return tags",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tags"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    }
  },
  "security": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "components": {
    "schemas": {
      "Articles": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Article"
            }
          },
          "cursors": {
            "type": "object",
            "$ref": "#/components/schemas/Cursors"
          }
        }
      },
      "Article": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "image_path": {
            "type": "string"
          },
          "read": {
            "type": "boolean"
          },
          "tags": {
            "type": "object",
            "$ref": "#/components/schemas/Tags"
          },
          "comment": {
            "type": "object",
            "$ref": "#/components/schemas/Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Comments": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Comment"
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "body": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Tags": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Tag"
        }
      },
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Cursors": {
        "type": "object",
        "properties": {
          "previous": {
            "type": "integer",
            "format": "int64"
          },
          "next": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    }
  }
}